fix: fix desync between player yaw and headset actual position
This should resolve a desync between actual headset angle and player yaw happening when player is dead or when a mod menu is active. The snapturn is still working during these states causing desync if used, so when the player input go back on, the actual player yaw is different from the one expected and instead of going forward the player will walk on a different direction.