feat: add new cvar vr_momentum_threshold to tune the sensitivity for keeping the momentum when changing programmatically the player velocity as during a recoil action.
This threshold is important when slidey movements are disabled to propertly move the player when the velocity is changed programmatically. If the change in velocity is lower than one, the threshold need to be set to a lower value than the default or the player will not move.
The cvar can be found in the experimental menu under the name "Keep momentum threshold"