[WIP] Pooled particles now sort-of working
It's ugly, it's hacky, but it's working! Stupidly inefficient at the moment as it has to iterate through all the particle pools *PER SUBSECTOR*. Also has issues rendering the particles at the correct size.
Needed to change HWSprite a bit so that we could use non-particle_t types. Particles really should be less coupled with HWSprite, but that's for future investigation.