Depoolifying ParticleDefinitions
Starting to work on making PooledParticles more like SelacoParticle, and turns out pooling the particles by definition worked against that goal. So renamed PooledParticles to DefinedParticles (Or ParticleDefinition, depending on the context), and removed the per-definition pooling behavior. Particle limits are now also defined in scripting, and respect r_particleIntensity.