Fixed some animation timing issues
Turns out we were skipping the first frame, and if you set a new animation in ThinkParticle it'd also skip the first frame. So made it so that it won't automatically update the frame for those cases (But WILL update the texture).

Also added a flag so we can disable animation on the particles.