Fixed issue where particles would stop spawning
Turns out that if the oldest particle was destroyed, it wouldn't update the index to the new oldest particle, meaning that if the particle system became full and it started reusing old particles, it would break the continuity of the particle iterators and only result in a handful of particles ever showing up again.

Note that this issue likely also exists in the p_effect particle system, although I didn't fix it because I don't want to go fixing things that nobody's reported an issue with.