Particle can now be made water-aware
Using the PDF_CHECKWATER flag. By default, particles don't check for water since it's a (Relatively) expensive test to do every tick, so we're only going to use it on certain particles.

Also added PDF_NOSPAWNUNDERWATER, which immediately destroys a particle if it's spawned underwater but has no tick cost.