Particles spawned underwater now check water state by default
We now only update the water state of a particle if it was spawned underwater. This avoids any water checking overhead for a vast majority of particles. As such, the new flag PDF_CHECKWATERSPAWN is now on by default.

However the PDF_CHECKWATER flag will still force the particle to check it's water state regardless of where it spawned.